feat: move UserModel to impulse app
