+++ apps/impulse/lib/main.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:helpwave_service/introduction.dart';
-import 'package:helpwave_service/user.dart';
 import 'package:helpwave_theme/constants.dart';
 import 'package:impulse/components/background_gradient.dart';
 import 'package:impulse/screens/onboarding_screen.dart';
 import 'package:impulse/screens/home_screen.dart';
 import 'package:impulse/theming/colors.dart';
 import 'package:provider/provider.dart';
+
+import 'notifiers/user_model.dart';
+
 
 void main() {
   runApp(
