+++ apps/impulse/pubspec.yaml
@@ -59,6 +59,8 @@ dependencies:
   carousel_slider: ^4.2.1
   flutter_svg: ^2.0.7
   dots_indicator: ^3.0.0
+  shared_preferences: ^2.0.15
+
 
 dev_dependencies:
   flutter_test:
