--- packages/helpwave_service/lib/user.dart
@@ -1,1 +1,0 @@
-export 'package:helpwave_service/src/user_model.dart';
