+++ packages/helpwave_service/pubspec.yaml
@@ -12,11 +12,6 @@ dependencies:
     sdk: flutter
   shared_preferences: ^2.0.15
 
-  impulse:
-    path: "../../apps/impulse"
-  helpwave_proto_dart: ^0.29.0-01c8073.0
-publish_to: none
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
